writer/requirements.txt
redis
cassandra-driver